# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: Secret Keeper Api
  description: An api that helps manage your secrets.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
tags:
  - name: credentials
    description: Manage your credentials
paths:
  /credentials:
    post:
      tags:
        - credentials
      summary: Creates a new credential
      operationId: newCredential
      requestBody:
        description: Credential that needs to be stored
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialRequest'
            example:
              credential:
                summary: Example of a credential
                value:
                  url: www.gmail.com
                  username: john.doe@gmail.com
                  password: password
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/definitions/CredentialDetails'
    get:
      tags:
        - credentials
      summary: Get credentials
      operationId: getCredentials
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CredentialDetails'
              examples: {}
            application/xml:
              schema:
                type: object
                properties: {}
      description: Get all the credentials that matches to research criteria.
      parameters:
        - schema:
            type: string
          in: query
          name: url
        - schema:
            type: string
          in: query
          name: username
    parameters: []
  '/credentials/{credential-id}':
    parameters:
      - schema:
          type: integer
          format: int64
        name: credential-id
        in: path
        required: true
    patch:
      tags:
        - credentials
      summary: Update a credential
      operationId: updateCredential
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialDetails'
        '404':
          description: Not Found
      description: Updates the credential
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCredentialData'
            examples:
              new-password:
                value:
                  password: changed-password
    delete:
      tags:
        - credentials
      summary: Delete a credential
      operationId: deleteCredential
      responses:
        '204':
          description: No Content
components:
  schemas:
    CredentialDetails:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        username:
          type: string
        password:
          type: string
    CredentialRequest:
      type: object
      properties:
        url:
          type: string
        username:
          type: string
        password:
          type: string
      required:
        - url
        - username
        - password
    NewCredentialData:
      title: NewCredentialData
      type: object
      x-examples: {}
      properties:
        newPassword:
          type: string
      description: The new credential data